feat(user): move Discord notifications panel to profile page
Extract DiscordNotificationsPanel; remove Notifications tab from Node Settings.
OAuth connect and Discord login redirects now land on /user.

Made-with: Cursor
